feat: add cron token webhook implementation
