+++ apps/www/src/api/functions/webhooks/index.ts
@@ -1,0 +1,5 @@
+import { purgeViewWindow } from "./purgeViewWindow";
+
+export const webhooks = {
+  purgeViewWindow,
+} as const;
