+++ apps/www/src/api/schemas/responses.ts
@@ -1,26 +1,26 @@
 import { z } from "zod";
 import type { InferResponses } from "../extended";
-import { schemaResponse } from "../extended";
+import { schemaResponse as restSchemaResponse } from "../extended";
 import { c } from "../root";
 import { httpErrorSchemaRo } from "./ros";
 
 // ========== Schemas ==========
 
-export const checkHealthSchemaResponses = schemaResponse({
+export const checkHealthSchemaResponses = restSchemaResponse({
   200: c.noBody(),
 });
 
-const fetchContentFromUpstreamSchemaResponses = schemaResponse({
+const fetchContentFromUpstreamSchemaResponses = restSchemaResponse({
   500: httpErrorSchemaRo,
   424: httpErrorSchemaRo,
 });
 
-export const fetchTextFromUpstreamSchemaResponses = schemaResponse({
+export const fetchTextFromUpstreamSchemaResponses = restSchemaResponse({
   200: c.otherResponse({ contentType: "text/plain", body: z.string().min(1) }),
   ...fetchContentFromUpstreamSchemaResponses,
 });
 
-export const fetchGpgFromUpstreamSchemaResponses = schemaResponse({
+export const fetchGpgFromUpstreamSchemaResponses = restSchemaResponse({
   200: c.otherResponse({
     contentType: "application/pgp-keys",
     body: z.string().min(1),
@@ -28,7 +28,7 @@ export const fetchGpgFromUpstreamSchemaResponses = schemaResponse({
   ...fetchContentFromUpstreamSchemaResponses,
 });
 
-export const purgeViewWindowWebhookSchemaResponses = schemaResponse({
+export const purgeViewWindowWebhookSchemaResponses = restSchemaResponse({
   200: c.noBody(),
   401: httpErrorSchemaRo,
   500: httpErrorSchemaRo,
