+++ turbo.json
@@ -65,6 +65,7 @@
     "KIT_API_KEY",
     "DATABASE_URL",
     "DIRECT_URL",
+    "CRON_TOKEN",
     "SENTRY_PROJECT",
     "SENTRY_AUTH_TOKEN",
     "NEXT_PUBLIC_WWW_URL",
